fix(pwa): use lazy loading to prevent build failure when PWA disabled
Split PWAContext into two files to use conditional lazy loading:
- PWAContext.tsx: exports PWAProvider wrapper, usePWA hook, and types
- PWAProviderInternal.tsx: contains the actual SW registration with
  the virtual:pwa-register import

This prevents Rollup from trying to resolve the virtual:pwa-register
import when PWA is disabled (PR preview builds), similar to how
ReloadPrompt handles this.
